requirements.txt
Flask
gunicorn
python-dotenv
pymongo
pandas
openpyxl
plotly
bcrypt
sendgrid
jsonify
matplotlib
xlsxwriter